+++ packages/clerk-js/src/ui/elements/PasswordInput.tsx
@@ -50,7 +50,9 @@ export const PasswordInput = forwardRef<HTMLInputElement, PasswordInputProps>((p
       onValidationWarning: message => setWarning(message),
       onValidationInfo: message => {
         // ref will be null when onFocus is triggered due to `autoFocus=true`
-        if (!inputRef.current) return;
+        if (!inputRef.current) {
+          return;
+        }
         const isElementFocused = inputRef.current === document.activeElement;
         if (isElementFocused) {
           setInfo(message);