+++ packages/elements/examples/nextjs/app/sign-in/[[...sign-in]]/page.tsx
@@ -175,6 +175,10 @@ export default function SignInPage() {
               <CustomProvider provider='google'>Continue with Google</CustomProvider>
             </div>
 
+            <SignIn.Passkey className='inline-flex px-7 py-3 justify-center transition rounded-lg focus:outline-none border items-center disabled:bg-[rgb(12,12,12)] focus:text-[rgb(255,255,255)] w-full duration-300 focus:!border-[rgb(37,37,37)] text-sm space-x-1.5 text-[rgb(160,160,160)] hover:text-[rgb(243,243,243)] disabled:text-[rgb(100,100,100)] select-none bg-[rgb(22,22,22)] hover:bg-[rgb(22,22,30)] border-[rgb(37,37,37)] hover:border-[rgb(50,50,50)]'>
+              <Clerk.Loading>{isLoading => (isLoading ? <Spinner /> : 'Use passkey instead')}</Clerk.Loading>
+            </SignIn.Passkey>
+
             {continueWithEmail ? (
               <>
                 <Clerk.Field
@@ -229,6 +233,13 @@ export default function SignInPage() {
 
           <SignIn.SupportedStrategy
             asChild
+            name='passkey'
+          >
+            <Button>use passkey</Button>
+          </SignIn.SupportedStrategy>
+
+          <SignIn.SupportedStrategy
+            asChild
             name='email_code'
           >
             <Button>Send a code to your email</Button>
@@ -278,6 +289,14 @@ export default function SignInPage() {
         <SignIn.Step name='verifications'>
           <div className='flex gap-6 flex-col'>
             <Clerk.GlobalError className='block text-red-400 font-mono' />
+
+            <SignIn.Strategy name='passkey'>
+              <P className='text-sm'>
+                Welcome back <SignIn.Salutation />!
+              </P>
+
+              <CustomSubmit>Continue with Passkey</CustomSubmit>
+            </SignIn.Strategy>
 
             <SignIn.Strategy name='password'>
               <P className='text-sm'>