+++ packages/nextjs/src/app-router/client/ClerkProvider.tsx
@@ -8,6 +8,7 @@ import { ClerkNextOptionsProvider } from '../../client-boundary/NextOptionsConte
 import type { NextClerkProviderProps } from '../../types';
 import { ClerkJSScript } from '../../utils/clerk-js-script';
 import { mergeNextClerkPropsWithEnv } from '../../utils/mergeNextClerkPropsWithEnv';
+import { invalidateCacheAction } from '../server-actions/invalidateCache';
 import { useAwaitablePush } from './useAwaitablePush';
 import { useAwaitableReplace } from './useAwaitableReplace';
 
@@ -56,7 +57,12 @@ export const ClientClerkProvider = (props: NextClerkProviderProps) => {
       return new Promise(res => {
         window.__clerk_internal_invalidateCachePromise = res;
         startTransition(() => {
-          router.refresh();
+          //@ts-expect-error next exitsts on window
+          if (window.next?.version && typeof window.next.version === 'string' && window.next.version.startsWith('13')) {
+            router.refresh();
+          } else {
+            invalidateCacheAction();
+          }
         });
       });
     };