+++ packages/tailwindcss-transformer/src/index.ts
@@ -50,6 +50,9 @@ function visitNode(node: recast.types.ASTNode, ctx: { styleCache: StyleCache }) 
       if (isLogicalExpression(path.parentPath.node) && !isRightmostOperand(path)) {
         return false;
       }
+      if (path.parentPath.node.type === 'ObjectProperty' && path.parentPath.node.key === path.node) {
+        return false;
+      }
       const cn = generateHashedClassName(path.node.value);
       ctx.styleCache.set(cn, path.node.value);
       path.node.value = cn;
@@ -101,7 +104,7 @@ export async function generateStylesheet(
   stylesheet += ctx.globalCss || '';
 
   for (const [cn, value] of styleCache) {
-    stylesheet += `html :where(.${cn}) { @apply ${value} }\n`;
+    stylesheet += `.${cn} { @apply ${value} }\n`;
   }
 
   const result = await postcss([tailwindcss(ctx.tailwindConfig) as Plugin, replaceVariableScope]).process(stylesheet, {