+++ packages/tanstack-start/.eslintrc.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+  root: true,
+  extends: ['@clerk/custom/browser', '@clerk/custom/typescript', '@clerk/custom/jest', '@clerk/custom/react'],
+  rules: {
+    'import/no-unresolved': ['error', { ignore: ['^#'] }],
+  },
+};