+++ packages/testing/src/common/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
 export * from './types';
 export * from './setup';
+export * from './errors';