+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @clerk/testing
 
+## 1.1.7
+
+### Patch Changes
+
+- Improve the error message when `setupClerkTestingToken` cannot find the FAPI URL ([#3520](https://github.com/clerk/javascript/pull/3520)) by [@anagstef](https://github.com/anagstef)
+
+- Updated dependencies [[`4ec3f63e2`](https://github.com/clerk/javascript/commit/4ec3f63e26d8d3725a7ba9bbf988a7776fe893ff)]:
+  - @clerk/shared@2.3.0
+
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [[`86a27f693`](https://github.com/clerk/javascript/commit/86a27f6933de50c99b6bc354bf87ff5c2cfcaf38), [`02bed2e00`](https://github.com/clerk/javascript/commit/02bed2e00d3e0a4e1bb1698b13267faf6aeb31b3), [`73e5d61e2`](https://github.com/clerk/javascript/commit/73e5d61e21ab3f77f3c8343bc63da0626466c7ac), [`b8e46328d`](https://github.com/clerk/javascript/commit/b8e46328da874859c4928f19f924219cd6520b11)]:
+  - @clerk/shared@2.2.2
+  - @clerk/types@4.6.0
+
 ## 1.1.5
 
 ### Patch Changes